ftrace: workaround for unexpected zero-padding in tracing pages (mostly 4.19 kernels)
We're seeing rare [voluntary] crashes due to misparsed binary tracing
pages, and this looks to be the main offender. Based on my understanding
of the ring buffer ABI, this looks like a kernel bug.

This patch adds a workaround to traced_probes to treat such pages as
successfully parsed (since there's no real data left anyway).

Bug: 204564312
Change-Id: Iaaba2e8ccf966add9948418c6adb7f1ba187379f