Add iid field to UnsymbolizedSourceLocation
The iid field must be present, and must be the first one (see
PacketSequenceStateGeneration::InternMessage for more details).
Otherwise, interning doesn't work in Trace Processor.

Generally speaking, this is an unsafe proto change. But I think it is OK in this specific situation because this message is not used anywhere yet.

Bug: b/220179588
Change-Id: Id3f6f80adf016e6cd3128dd70fa64b0cd4208b3e