UI: fix modal re-appearing
The issue was the following:
 - When we dismiss the modal dialog, the dismissed state
   stays local to the Modal component (the outer container, which
   may or might not display a ModalImpl depending on the closed
   state).
 - When switching pages, because the modalContainer is one of the N
   entries in the pages array, the Mithril diff engine ends up
   destroying and recreating the Modal component. This clears up the
   "closed" state and causes the dialog to show up again.

The fix here is to persist the closed state at the ModalContainer
level, so even if the Modal component is re-created, we remember
and propagate down its closed state.

Arguably we should change the array in page.ts to have keys, to
avoid this subtle behaviour. BUt that might have more side effects
and I don't feel adventurous enough to do it right now.

Bug: 235335853
Change-Id: Icca68ebe3a98d8a29efcb9cf292d205025b3adb4