[ui] Do not export annotation tracks to uiTrackIdByTraceTrackId.
Currently annotation_slice tracks and regular slice tracks share ids,
which leads to UI selecting incorrect track e.g. when searching or
following the flow.

For now, given that annotation slices can't be searched and do not
have flows, do not add them to uiTrackIdByTraceTrackId.

In the future, "trace track id" probably needs to become a tuple
of (namespace, track_id) instead of a single int.

R=hjd@google.com
BUG=b/235325363

Change-Id: I76c124016354fd437f0e0e6073a3dd7f5756dfa7