Shared library API to serialize protobuf messages
This provides a basic API to serialize protobuf messages with field and
nested messages on top of the PerfettoStreamWriter ABI.

The interface is generic and it requires to pass field numbers
explicitly. In the future we might want to add an interface (on top of
this) that knows about the messages type (similar to protobuf generated
code).

Bug: 237053538
Change-Id: I3608d8801838426accaa5be5f12e2b2416f47645