[ui] Show flow events in the "Current selection" panel
To minimise the amount of clicks one needs to do during an
investigation, show the flow events attached to a given slice when
selecting it.

Also to make it more useful for common Chrome flow events (post/run
tasks and mojo messages), show a detailed name for them (adding their
arguments to their name).

R=hjd@google.com,ddrone@google.com
BUG=b/235074138,b/224527755

Change-Id: I7ddf9aa37cde73251979a9d174c1581e4c497143