perfetto: Only match partial rss_stat_throttled trigger
The kernel shows all the default values of a trigger when read
from and trace event 'trigger' file, e.g:

    echo 'hist:keys=mm_id,member:bucket=size/0x80000\
    :onchange($bucket).rss_stat_throttled(mm_id,curr,member,size)' \
	     > kmem/rss_stat/trigger

    cat 'kmem/rss_stat/trigger'

    hist:keys=mm_id,member:vals=hitcount:bucket=size/524288:\
    sort=hitcount:size=2048:onchange($bucket)\
    .rss_stat_throttled(mm_id,curr,member,size) [active]

For this reason, only do a partial match on the generated
event name (rss_stat_throttled) to detect if the trigger is
present.

Trying to match the complete trigger string is prone to fail
if, in the future, the kernel changes default fields or values
for event triggers.

Test: adb push test/configs/rss_stat.cfg /data/local/tmp/
        && cf-adb shell 'cat /data/local/tmp/rss_stat.cfg |
        perfetto --txt -c - -o /data/misc/perfetto-traces/trace
Bug: 236172293
Change-Id: Ifa50c9e08893c8494739747fe77ab1ceec5c0dfd