[ui] Add "visualise arg values" option to the arg context menu
When clicked, Perfetto UI will show additional tracks - one track
for each slice tracks containing a slice with a given argument.
The duration and timestamp will be the same, while the name will
correspond to the value of the argument.

This feature will allow us to avoid emitting duplicated slices for the
things that should have really been an argument on an existing event
(i.e. task priority and task queue name in the Chrome Scheduler),
while retaining the benefit of being able to see the distribution
of the values at a glance (and e.g. spot the outliers by their colour).

R=hjd@google.com,ddrone@google.com
BUG=b/233863459

Change-Id: I44bdf1c2ca70bdb9c79e8fedcb2817bce1a40d41