tp: split off ColumnStorageOverlay from RowMap
RowMap is a data structure which we were using directly as an overlay
for ColumnStorage. This worked while filtering with simple but as the
complexity as grown, using very generic terms "row" and "index" just
creates confusion when reasoning about complicated code.

Take the first step to improving this by creating a new class called
ColumnStorageOverlay which now owns the complicated filtering logic
used by Tables.

Over time, we'll continue simplifying RowMap to only contain very
primitive operations and build the rest on top in other data structures

This is a pure refactor and so there should be no behavioural change
as a result of this CL.

Bug: 235104800
Change-Id: Id1970f0c0e59954fd99ff1cdee55dce4227b6b7d