tracing: Add API to shut down Perfetto
This patch adds an API call to shut down the Perfetto SDK and release
any OS-level resources (threads, files, sockets, etc.). Note that it's
not possible to reinitialize Perfetto again in the same process after
shutdown. Instead, this call is meant for ensuring no Perfetto-related
threads are running so that the entire module using Perfetto can be
unloaded with dlclose().

(Cherry-pick of https://android-review.googlesource.com/c/platform/external/perfetto/+/2146891)

Bug: b/237740008
Test: `adb shell cmd gpu vkjson` => check that virgl_render_server doesn't crash
Change-Id: If82163f12c2eb426dae9cc99ccd9eb3b25bae741