trace_processor: show I2C transactions for the outermost adapter IDs
The new I2C slice track provides a "high-level" overview of I2C
operations performed by a process.
It is called "high-level", because I2C operations are organized as
nested transactions, and can involve multiple logical/physical adapter
numbers depending on the I2C topology of the system.
For example, when a program performs the following operation using the
userspace i2c_transfer API:

  "Write 0x00 and 0x05 to i2c-90"

And the system has the following I2C topology:

  root --> i2c-7 --> (Mux) --> i2c-86 --> (Mux) --> i2c-90

The following I2C trace would appear:

  i2c_write : i2c-90 #0  a=04c f=0000 l=2 [00-05]
  i2c_write : i2c-86 #0  a=070 f=0000 l=1 [08]
  i2c_write : i2c-7  #0  a=070 f=0000 l=1 [08]
  i2c_result: i2c-7  n=1 ret=1
  i2c_result: i2c-86 n=1 ret=1
  i2c_write : i2c-86 #0  a=04c f=0000 l=2 [00-05]
  i2c_write : i2c-7  #0  a=04c f=0000 l=2 [00-05]
  i2c_result: i2c-7  n=1 ret=1
  i2c_result: i2c-86 n=1 ret=1
  i2c_write : i2c-86 #0  a=070 f=0000 l=1 [00]
  i2c_write : i2c-7  #0  a=070 f=0000 l=1 [00]
  i2c_result: i2c-7  n=1 ret=1
  i2c_result: i2c-86 n=1 ret=1
  i2c_result: i2c-90 n=1 ret=1

In the above example, i2c-7 is the physical I2C adapter, while i2c-86
and i2c-90 are logical ones. When kernel sees i2c-90, it uses its
knowledge on the I2C topology to automatically generate the commands for
handling the mux operations and perform them over i2c-7.

For this change, only a single slice with name "i2c-90" is shown, since
this is enough for giving a high-level overview of the behavior of a
process.

In contrast, the I2C topology is system-wide, and is not only limited to
one process. As such we think showing the outermost I2C adapter ID is a
good fit for a per-process slice track.

Tested by manually examining the output of trace_processor_shell and
comparing it against the slices shown on the visualizer.

Change-Id: Ic5278e9c2ef38303a98629dce5db85672bff065b