ui/tp: fix flamegraphs and tracks for stackless perf samples
The perf_sample table is allowed to have entries without callstacks, for
example because the config is for a counter-only trace. However the
flamegraph implementation and related UI elements assume that every
perf_sample row is a callstack sample, which causes both crashes and
incorrect flamegraph output (logic ends up picking up the 0th callsite
as equivalent to the null one or something). This patch filters out such
samples from both the flamegraph implementation and the UI tracks.

Additionally, change the track name from "Perf Sample" to "Callstacks
1234", including the pid of the process. This makes it easier to
distinguish tracks when they're pinned, and helps with kthread tracks as
they're all in one synthetic track group, but have distinct pids.

Bug: 229776518
Change-Id: I552342570563c6f268f6528bac5d48a03048bbf6