Add android_jank_cuj metric
It's a rewrite of android_sysui_cuj metric with support for processing
>1 CUJ per trace. At the same time the queries were cleaned up and
simplified (e.g. matching of slices to frames), I added some docs to
make it possible for a motivated person to actually follow the flow.

Creating a separate metric to not break downstream consumers and to
allow running both metrics side by side to compare results while we are
migrating from one to the other.

SF tables will be rewritten in a similar manner in a future CL.

Change-Id: I5c341ab4a4a6439cc11f286b29d12624d5938222