Add new arg for Chrome processes
This will be used for identifying Chrome processes and threads in system traces more reliably.

We cannot do the same for Chrome threads (at least not straight away) - at the moment, thread table doesn't have arg_set_id (= threads don't have args). This should be OK - for each thread, we know which processes it's a part of. We can say "a thread is a Chrome thread iff it's a part of a Chrome process".

Bug: b/239675998
Change-Id: If98f75350f9614e866a0a5868f4c64c9e517dc25