Add ability to trace only select syscalls.
On android, the raw_syscalls tracepoints are used to trace syscalls.
These however include *all* syscalls, though most users are probably
only interested in a subset.

This change adds the ability to specify the set of syscall events you
want to track and ignore the rest. It works using ftrace filters and
uses proper muxing to handle multiple configs. If no filter is provided,
the behavior is the exact same as it was before.

When muxing, the union of all necessary syscalls will be enabled. Since
the individual probers don't filter, this means that concurrent traces
can give more syscalls than you requested, but never less.

Test: ran perfetto_unittests
Change-Id: I1478cbd28625f3b4c5c2d050df8b6fed72db993c