proto: perf: process sharding config and packet defaults
This option directs the callstack sampling to choose only a subset of
seen processes for unwinding, but without explicitly enumerating the
subset like we'd do with target_pid or target_cmdline. The intent is to
lessen the overheads of systemwide callstack sampling.

The allow/deny choice for a process is "pid % process_shard_count == X",
where process_shard_count is set by the config, and X is chosen at
random for the duration of a trace.

Doc: go/tperf-sharding
Bug: 223773242
Change-Id: Idbadf707bbab8b4aec0c5c4d22d352ee471642ad