TP: Fix windows build and small time bug in ZipReader
Fix a compiler issue on Windows, which does not have
gmtime but has an equivalent _mkgmtime.
Also fix a wrong calculation of the seconds in ZipReader.
The DOS file format has only 5 bits (0..31) for seconds.
Users are expected to shift it by one (effectively the
DOS format loses 1 second of precision).

Bug: 240159066
Change-Id: Ic431af0009a552fce9607cb17ffed8236d197747