Startup tracing: Enable a test
Also fix the underlying issues. Before this change, this test
(MainTracingNeverStarted) was causing a crash in the next test after it.

The problem was:

The next test after MainTracingNeverStarted had the same trace-writer
as `MainTracingNeverStarted` test (i.e. same pointer value).. and the
next test was attempting to call it's destructor ( ds_tls.Reset() )
after realising it need to recreate a new trace-writer, and that
destructor was resulting into crash (some random illegal memory access).

Note that `Muxer::ResetForTest` doesn't destroy active trace writters,
because it's practically difficult. (Because trace writters are stored
in TLS of every thread).

However `Muxer::ResetForTest` ensures to not destroy the producer if
their SMA (SharedMemoryArbiter) is still being referenced by some
trace writers.

Before this change, `Muxer::ResetForTest` was not able to prevent the
procucer from getting destroyed because it was not connected to the
service, because for starting the startup session, we disconnect and
connect again. Reconnection takes sometime but
SetupStartupTracingBlocking return much earlier than reconnection.
It returns when data sources are ready for tracing. As a result, when
`Muxer::ResetForTest` was calld in test TearDown step, it thought that
producer was not connectd to the service, hence it was safe to destroy
it, because before StartupTracing, a producer could have
SharedMemoryArbiter only if its connectd to the service.

StartupTracing changes that assumption. A producer can have
SharedMemoryArbiter even if it is not connected to the service.

Hence `Muxer::ResetForTest` was destroying a SharedMemoryArbiter with
active trace-writers. As a result when trace-writer was destroyed
in the next test, it was making `Commit` call on a already destroyed
SharedMemoryArbiter. Which was resulting in crash.

Also when reconnection happens to the producer via service transport
later, it will be referencing a deleted producer. This become another
source of crash.

We fix that in this CL, by making sure we sync producer in the TearDown
step of PerfettoStartupTracingApiTest. `sync producer` is blocking call
that waits until producer is connectd to the service.

Change-Id: I46037e6896bf4765baded55392a4ca7904989024