ui: Don't aggregate perf samples on single selection
Currently when a single perf sample is selected all callstacks
before it in the same track will be aggregated. This is a holdover
in behavior from the heap profiling where this was actually desired.

Instead make it so that when a user clicks on a single perf sample
only the callstack for that single sample is shown in the
flamegraph.

In addition when a trace is now loaded, we will now select all
samples in the default track instead of just one. This is done so
that despite the above behavior change, we will still display a
full flame graph when the trace is opened instead of just the stack
for the selected sample.

Testing:

1. Open a trace with perf samples
2. verify upon opening we see the same flamegraph as without this
 commit.
3. click on the last sample of the trace.
4. verify we see only a single a single callstack in the flamegraph

Change-Id: Ia34bd861d2d6bb6b0ee0301475cb54a539ce9ff0