[ui] Update focus logic to change zooms levels more conservatively
The current focus logic changes zoom level each time a new slice is
selected. This leads to very hectic experience when trying to follow
the flow or go over the search results - the zoom levels change very
quickly and it's hard to understand which part of the trace you are
now looking. The experience particularly bad for short or instant
events (like postTask), where the current logic zooms you all the way
in.

This patch changes this logic to preserve the current zoom level
wherever possible according to the following rules:
a) If the range-to-be-focused can't comfortably fit within current
  visible range (more than 50% of visible range duration), update
  the zoom range to be 5x of the to-be-focused range duration.
b) Calculate the new visible range, centered on the to-be-focused
  range, but preserving to the zoom level (including snapping to the
  trace boundaries). If the calculated range is the same as the current
  visible range, update the zoom range (same as in a)).
c) Otherwise set the new visible range to be as calculated.

Rule b) ensures that double-tapping 'f' preserves the current behaviour:
first tap centeres the visible range, while the second updates the zoom
level.

R=hjd@google.com,ddrone@google.com

Change-Id: I1153d14dde5952df6b9b68062f98a760f93b9693