ftrace: Include arguments of syscalls
When ftrace events get enabled for sys_enter and sys_exit
only the id and ret fields will get recorded. This is
because also including the arguments adds quite a bit of
data on top.

For certain trace analysis it might still prove as useful
to have these arguments, e.g. when trying to identify the
correlation of threads delayed by long lasting futex calls
or when trying to identify a certain misbehaving ioctl in
a series of ioctls issued by a task.

In order to manage the overhead of all of this, inclusion
of syscall args remains disabled by default but can be
enabled by using a new option in ftrace_config named
'include_syscall_args'.

Change-Id: I67e8fe55934228a1621b39bf8e32645615dbde10