Fix `_COPY` version of legacy trace macros
For most legacy macros (eg: `TRACE_EVENT_MARK_WITH_TIMESTAMP0`), we
have a `_COPY` version as well.
(eg: `TRACE_EVENT_COPY_MARK_WITH_TIMESTAMP`). The implementation of
`_COPY` macro was buggy before this change. We were passing
`TRACE_EVENT_FLAG_COPY` flag but this flag was never used.

An easiest way to reproduce it was:

```
char ptr[] = "XYZ";
TRACE_EVENT_COPY_MARK_WITH_TIMESTAMP("cat", ptr, MyTimestamp{0});
ptr[0] = 'W';
TRACE_EVENT_COPY_MARK_WITH_TIMESTAMP("cat", ptr, MyTimestamp{0});
```

It was recording `XYZ` event name at both trace point.

The reason being, perfetto was not coping string-value. Similar to
usual case (where event name is assumed to be a static string),
perfetto was caching pointer values and if a event-name-pointer matches
an existing pointer, it ASSUMES the string-value of new pointer is same
as string-value of the cached pointer when it was cached. Hence it
assign the same intern-id to the second event.

In this fix we deprecate `TRACE_EVENT_FLAG_COPY`.
we plan to remove any direct uses of this flag and will remove it
from chromium if that is possible, and then remove it.

Anything using `TRACE_EVENT_FLAG_COPY` flag would be buggy
anyway.

The fix - we wrap the event name in `::perfetto::DynamicString`.
The event name wrapped in ::perfetto::DynamicString would be copied
correctly.

Change-Id: If0af9e318a09830f079efabbfb7ab9b43ebf0d41