[fuchsia] Reland: Adds a callback to send a shmem FD to the client.
Reland updates function signatures in host_impl_unittest.cc and
ipc_integrationtest.cc so that `perfetto_unittests` builds on Chrome.

Original CL description:
Adds a std::function callback to AdoptConnectedSocket_Fuchsia,
which is provided by the service embedder to handle the transport
level details of sending the shared memory to the client.

Bug: 1166873
Change-Id: I73fd04c59bda73120e0c2aafdf98a087a838ccf7