Enable Building in Fuchsia Tree
Perfetto already builds for Fuchsia as part of Chromium with
"build_with_chromium". However, we'd also like to build Fuchsia
components that use Perfetto that are not Chromium. Perfetto is
currently rolled into the Fuchsia tree and this change allows it to
build as part of said tree and link against
libperfetto_client_experimental.

This uses platform_posix.cc on Fuchsia to provide some platform specific
implementation. Fuchsia is posix compatible enough that we can build and
link against platform_posix.cc

Change-Id: I02bfb24a8e8e61428f3df5d63c93443a75de977b