Replace #apex with #systemapi
Currently, for an API symbol to be accessible to APEXes, it needs to be
marked as either # systemapi or # apex. It was originally just # apex,
but we added # systemapi to clearly identify the origin of the APIs.

The intended use is

* #apex is for APEX-visible symbols that are defined in an APEX
* #systemapi is for APEX-visible symbols that are defined in the
  platform (the non-updatable part)

This intention is documented build/soong/docs/map_files.md, but isn't
enforced at all.

With b/239274367, this is now enforced and therefore the #apex tags in
the platform library are replaced with `#systemapi`

This change does not alter any functionality.

Bug: 239274367
Test: m
Change-Id: Ide2b18740bc338930ab3e8726c9fdf6f66b9d2a2