traced_probes: support multiple batteries on Linux and CrOS
This change introduces "named battery" to the linux.sysfs_power data
source: when there are multiple batteries observed by the data source,
it emits the battery name from the dirname of /sys/class/power_supply/
subdirs.

When the trace processor sees a named battery, it adds the battery
name to the track name like batt.BAT0.charge_uah for the charge counter
of battery "BAT0".

Bug: 242801328
Change-Id: I20f4090b0e2ed15a2831d53ecf5eeae9b17431d6