tp: don't append cell types when zero rows
The protobuf spec mandates that zero sized packaged repeared fields are
not included in the serialized structure. While libprotobuf seems to be
fine with bending this run, Wire (Java/Kotlin protobuf library) strictly
adheres to the spec and so errors.

Change to not emit cell types if there are no rows. Do the same for
blobs for consistency.

Bug: 243735331
Change-Id: Ia28455bc2e802ea0c265d4eb2ea6174f4ff9040f