ui: Re-enable events for analytics
Prior, gtagGlobals.dataLayer had objects as elements.
Then `arguments` was changed to rest paramters, which are
arrays.
After the change, gtagGlobals.dataLayer started having arrays as elements.
Since this change, Google Analytics stopped showing datapoints.

Example of the previous version to which this CL is restoring:
```
{"0":"event","1":"page_view","2":{"page_path":"/record","page_title":"no-page-title"}}
```

Example of the existing version which is not picked up by analytics:
```
["event","page_view",{"page_path":"/record","page_title":"no-page-title"}]
```

Change-Id: Ieec9d1744b71aac3a8f415112855852bf9c984aa