Add and emit first_packet_on_sequence field to TracePacket
This field is set by TraceWriterImpl for its first packet, and later will be used in trace processor to identify to identify sequences/threads that began in the middle of a trace file's retained timeline (for cases when we don't have thread start timestamp). This can help to decide whether a thread that appears in the middle of the timeline is a newly started thread or a thread with data loss, and will be used to identify reliable timeline ranges.


Bug: b/239830951
Change-Id: I272e0c8b7e0d43d7811a037412cfb77d6d6fe382