UI: modal: fix off-by-one in modal dialog attrs.
Before this CL the modal dialog attrs were off by one redraw
cycle. This caused glitches when updating dynamically the contents
of the modal dialog.
The problem was the following:
1. pages.ts invokes fullscreenModalContainer.render() in its view()
   method.
2. render() passes down the current attrs (let's say attrsV1) to the
   m(Modal, attrsV1) invocation.
3. However, at that point, the rest of the page has not been processed
   (i.e. Sidebar::view(), Topbar::view() have not been called).
   This is because mithril acts in several passes, where it first
   processes the m() invocations, and only later deals with lifecycle
   methods.
4. Later we'd get to the actual page component view() method, which
   would in turn invoke fullscreenModalContainer.updateVdom(attrsV2).
5. But this is too late, because render() has passed already the
   attrsV1 to Modal.
6. Later the Modal::view() method is invoked, but with attrsV1.

This CL fixes by hoising the container mithril component as a member
field of ModalComponent, and passing down the attrs only in the view()
method (which is called later), and not at m() time like render() used
to do.

Test: manual: addedd logging.
Change-Id: Ic3f0871f07f62cd1aae6fedb60a4d1401761dada