tools: Fix Hermetic ADB Path
Due to the leading '/' in the hermetic adb path, python will interpert
this as an absolute path. The underlying os.path.join() will not behave
as expected if you try to concatinate this with the repo root, since
both are seen as absolute paths.

Instead, drop the leading slash to make it clear to python that this
is a relative path, and not an absolute path.

Change-Id: I446ab529207dd39b98755b7836aa0636fed0cb4c