ui: Check if the trace has datasources before displaying 'Record'
For Chrome, only a subset of probes is shown. However, because
the recordConfig is not separated by probes, the UI allows
starting a tracing session.

Also for Android, the UI allows starting an tracing session.

This change checks that the trace has any data sources and if it
does not have, recording will not be allowed and a warning message
will be displayed.

Bug: 237510542
Change-Id: Ib30b9d745556230f9c08f9a1d87be2d26ab1ec0a