Support async stop in TrackEvent datasource
There's a mechanism for Perfetto data source to request a deferred stop
in OnStop callback. Currently it's not supported in the TrackEvent
data source. TrackEvent immediately disables all tracing categories when
OnStop() is called.

This CL adds this support, making it possible to call
HandleStopAsynchronously() from TrackEventSessionObserver::OnStop() to
defer disabling of tracing categories.

Change-Id: I6e162c6b6b14142bf2ff3f87d99c56bcdd912d2c