tp: implement native support for the glob operator
This CL adds a class to parse and match UNIX globs inside trace
processor. This can be used to implement native support for SQL globs,
significantly improving the performance of real world trace
processor queries.

This CL does not perform the integration with filtering yet: that will
be done in a follow-up to allow easy reverts if necessary.

Bug: 249961301
Change-Id: I581a1a8ae615113082ecff79f3d488ac397d18b6