tp: switch to using GLOB matching code
This CL switches the trace processor filtering code to use the
GlobMatcher class which was recently introduced instead of relying on
SQLite.

This massively speeds up a lot of queries in trace processor. For
example, the startup metric on a 1.1GB trace reduces from 45s to 21s
from this change alone.

Bug: 249961301
Change-Id: Iac3c9f743d359afb19cd09f9d679e74dd66faf3c