Simple query history
The CL implements a simple list of previously entered queries (limited
only to queries entered in standalone query page for now). The list is
limited to 50 last queries; and it's possible to either re-run them or
delete them from history. The queries can be "favorited", which puts
them at the beginning of the list in the UI and makes them not to be
counted towards 50 queries limit.

Screenshot: https://screenshot.googleplex.com/6JeL8Afyj6k2CkG

Bug: b/203297197
Change-Id: Ie48f4be9ed66f2618bb044257e78698394d5b638