tp: tidy up amalgamated SQL header generation
This CL changes the SQL amalgamation system to generate a header
directly which can be depended upon instead of requiring manual
rule.

We do this in an almost identical r way to cc_proto_descriptor
generation where we have a Bazel macro which does both the genrule and
defines the cc_library.

Bug: 255535171
Change-Id: I103ce8953a685c20f9b5a1c6d226d8f42f94c425