Revert "Remove manual implementation of scroll to slice in query result table"
This reverts commit 58ba9899d46a0930ae737d50b5baea0bc47cc02a.

Reason for revert: See discussion in Perfetto chat:

Omar Elmekkawy, 18 min
There is a change recently on Autopush that redirects you to a "current selection" tab when you click on a slice in the query results, instead of redirecting you to the slice on the UI, which made debugging much harder for me.

@Alexander Timin

Hector Dearman, 14 min
@Andrew Shulaev aosp/2300920 maybe

bit surprised though, I would have expected it to still scroll to the slice

but maybe also switch the current tab which I can imagine might be annoying in some cases

Omar Elmekkawy, 2 min
I would push for putting those features behind a flag at least if you want to keep them.

Hector Dearman, 2 min
Which features?

Omar Elmekkawy, 1 min
* Clicking on the table jumps you to a different tab
* Clicking on the slice doesn't navigate to the slice (guess this is more of a bug?)

Hector Dearman, Now
Yeah I mean I think the expectation of that CL was that is was a no-op refactoring, to the extent it wasn't that is probably a bug.

Ah Andrew is OOO this week, will revert for now in that case

Change-Id: I423fa71fd89bed6dc0420ee1c8a0acb81301553c