tp: Add table holding file descriptors
Parse file descriptor messages collected as part of
the process stats message in a trace file and store
them in a new table mapping each fd to its parent
process and path.

This can be used in custom queries to e.g. check
which files had been opened by which process or
what filedescriptor a long lasting syscall was
operating on.

Change-Id: If5bf8e9e084e7f2520512c3be3ce741dd63c95d9