ui: infinite loop when searching
When using toopbar to search, for example, searching for "animating", a total of 10 results are found. When jumping to the 10th result, executing setToNext again will not be able to jump to the first result.

Similarly, when jumping When the first result is reached, execute setToPrevious and cannot jump to the 10th result.

This CL is to solve that problem. When jumping to the 10th result, execute setToNext again to jump to the first result, and when jumping to the first result, execute setToPrevious to jump to the 10th result

Change-Id: Ic89936b5c40484707e05095b5e4638df811009fa