tp: fix is null/is not null constraints on view functions
SQLite expects primary keys to be unique and non-null. We were not
adhering to this with view functions as we were just using the set of
all return values as the primary key but return values were allowed to
be null

Fix this by having a separate row counter which is always unique and
non-null.

Bug: 260820488
Change-Id: I40f8ddba3f2cfdc712f60bf66248c37e98dc655e