tracing: Add init option to avoid linking system consumer
Today when using perfetto::Tracing::Initialize() with the system
backend, both the producer and the consumer side of the ipcs are linked
in, even though most perfetto users will not likely use the consumer
side.

The consumer side implementation is responsible for a large portion of
the statically linked binary size (~100Kbytes out of ~700kBytes, thanks
rzuklie@google.com for finding that out!).

This commit adds an option to perfetto::Tracing::Initialize to avoid
initializing the consumer side of the system backend.

Change-Id: I17a72f4864c6c300a6a734f93b349e6d020f1967