fix lwis/tracing_mark_write proto gen
aosp/2352609 was not submitted according to tools/run_ftrace_proto_gen and tools/gen_all output, so it's both inconsistent between protos and will be undone by the scripts.

The event needs a bit of special casing since the event name by itself is not unique.

The fact that CI/git-cl presubmit don't catch this is a separate issue)

Bug: 258115890
Bug: 259191638
Change-Id: Idafbfcc9da64e4047198eaaa673164443b25f396