Refactor chrome task delaying input processing to improve performance.
Calculating tasks delaying input processing is slow in part due to
recalculated chrome_tasks and other views/tables. Refactor out the
common functionality so that it is only called once for long tasks
and once for standard metrics.

This doesn't fix the full performance issue with the diff tests so
more work is forthcoming.

Change-Id: If624f8308d10afdabe110e94404c99a6403724f9