[TrackEvent] Make sure category search is compile time
Constexpr function are not guaranteed to be evaluated in compile time
when their result is used in non-constant expression. Hence usages
of TRACE_EVENT_CATEGORY_ENABLED macros often resulted in runtime
evaluation of TrackEventCategoryRegistry::Find().

This CL ensures that the results of constexpr functions are constexpr
themselves which guarantees they will be evaluated in compile time.

Bug: b/265910459
Change-Id: Ia492807bffa196ce6e73209912e4570a9edc1267