base: add ability to override platform-level functionality in base
In Google3, we need the ability to annotate the regions around blocking
syscalls so the userspace scheduler is aware of them. Because of this,
add functionality to mark such regions by having an abstraction of a
"platform" at a build system level and allow Bazel to override it.

Moreover, annotate read, write and poll syscalls in base with these
new functions: the "default" functions are empty noops.

Change-Id: I18c8246d32bac512494c2bc2cd5b02d840b82ec6