[tp] Treat missing Browser and Renderer main threads as data loss
If the trace has a Browser/Render process without a main thread, the thread data got lost, most likely due to the buffer overwrite. In this case, the entire trace cannot be trusted.

I've tested this metric on a few real life traces. It doesn't change the metric output for traces where all Browser/Renderer mains are present, and shows empty reliable range for traces with a lost Browser/Renderer main thread.


Bug: b/266072371
Change-Id: I400ebc0b67226cbd72cc04b078e17432147df5e0