tp: annotated callstacks: add "common-frame-interp" and cover more mapping names
Mappings:
* cover JIT code shared with the Zygote
* cover .dex extension as it's appearing again in apex modules
* attribute unadorned .apk to interpreted as libunwindstack should
  append a "!lib.so" suffix in the case of an embedded ELF.

Annotations:
We use the "common-frame" annotation for function frames that are
uninteresting for most users of flamegraph profile visualisations. At
the moment this is Android-specific and means ART implementation details
such as interpreter/jni bridges).

One usecase for callstack info is to try to classify the entire sample
by the leafmost frame. Example: leafmost frame is in an .oat, so we can
say the process is executing ahead-of-time compiled managed code at the
time of the sample. It's currently difficult to detect interpreter mode
callstacks, as the leafmost frame will be the libart.so native
implementation of the interpreter itself.

To resolve this, add a "common-frame-interp" annotation that tags
libart.so with specific function name prefixes (nterp_/Nterp), as that
is a sufficiently stable convention in ART. Callstacks with such an
annotation on a leafmost frame should be treated as interpreted. The
name is intentionally chosen to allow users to continue filtering out
ART details by excluding anything with the "common-frame" substring.

Misc: still using strncmp as I can't figure out a neat and readable way
of doing compile-time string lengths / string views here. cpp14+ might
help here.

Bug: 191856355
Change-Id: I5a75c14842dd6a6005d9185dc12e30ed521e405d