use memcmp in StringView::StartsWith/EndsWith
Current version results in somewhat surprising codegen in that it's not
at all like memcmp. When the prefix is a compile time constant, shorter
strings emit per-byte comparisons ("cmpb+jne" pairs of instructions on
x86). Longer strings change to something more regular, but at no point
does it seem to vectorise. This happens with clang
on -O2, -O3, x64 and arm64.

We suspect it's due to the compiler having to be conservative with accessing memory after the point of inequality. As in, the compiler doesn't see that the entire block of memory is safe to access.

Imo memcmp is aligned with what this function is doing, so we should
just use that.

Change-Id: I2e47ef036691ea67fc57f7cb01541c19690169f1