Avoid data race in the access to tracing muxer instance
Some Chromium TSAN tests report a potential data race when accessing
the tracing muxer instance. It can happen because there are no memory
fences on the tracing fast path and so DataSource::TraceWithInstances
is not guaranteed to see the updated value of instance_.

It should be extremely unlikely outside of tests, but we can fix it
anyway. If we don't see the TLS state initialized, we acquire a fence
to make sure that the thread have observed the initialization of the
muxer.

Bug: crbug.com/1410450
Change-Id: I0127d14bf7d700f76322425c5c36d583c84cb214