ftrace: fully respect max_pages in CpuReader::ReadCycle
If max_pages is not a multiple of parsing_buf_size_pages (=32), then we
might over-read since we step by that amount. That in turn violates an
expectation that pages_read <= max_pages in ftrace_controller.

This isn't a problem in practice since we don't normally hit these
max_pages limits. And even if we do, the outcome is that the ftrace read
tick reads fewer pages out of the kernel buffer than it otherwise would,
but the next tick onwards is good again.

Change-Id: I4cc6e38afeeaa59b624c8414528a1c6525da06b5