traced: add perfer_suspend_clock_for_duration option
By default traced uses CLOCK_MONOTONIC for the duration_ms
timer. That comes all the way down from poll() used by
PostDelayedTask(). In some cases, this is unideal as CLOCK_MONO
isn't suspend-aware. On mobile phones a trace with a 24h duration
might end up lasting weeks if suspend gets in the way, causing
consequent trace losses if the devce is rebooted.
This CL adds an option to use CLOCK_BOOTTIME for duration_ms.
When enabled, it uses PeriodicTask (without the "periodic" bit)
which under the hoods uses timerfd.

Bug: 269589525
Test: manual:
 - Start a config with perfer_suspend_clock_for_duration:true
 - Force the workstation to suspend (echo mem | sudo tee /sys/power/state)
 - Wait for longer than duration_ms
 - Observe that the tracing session is stopped immediately after resume.

Change-Id: Ic05db218596f0a21d7c440eb5b96a56529c8a909