tp: add TokenizedObjectStore for storing metadata while sorting events
This CL introduces TokenizedObjectStore, a variadic data structure
optimized for storing the metadata for events in a highly memory
efficient manner. This class is intended to be a replacement for
VariadicQueue and related classes due to being better tested and more
readable.

This class is implemented by building on top of BumpAllocator from a
previous CL and a follow-up CL will move TraceSorter to using this
logic.

Change-Id: I40302a7d23ace01880e155834ccfb3b665caca38