trace_processor: Fix build on windows
Bit-fields on MSVC (and therefore clang-cl) behave differently (the
behavior in implementation defined after all) and result in a different
object size.

By reusing the same type for all the fields, it seems like MSVC behaves
like GCC.

In order for this code to be truly portable we should avoid bit-fields
or relax the static assertion, but this fixes the build in the meantime.

Bug: 270301961
Change-Id: I2772cd1338ba911db4b7797d0cc0bc3c91ee596a