ui: Remove redux from pivot table code
Previously there were two versions of the pivot table in the code base.
The original one (pivot_table) and the new one (pivot_table_redux).
For a while now only the new one (pivot_table_redux) has existed.
The `redux` in the name now isn't needed and is a source of potential
confusion and overly long names, remove it renaming everything to do
with the pivot table simply 'pivot_table'.

Pure noop rename.

Change-Id: I44076dc1eb53365eee82036d35034a627f8ea477