tp: add stats entry for likely-broken frametimeline end_frame events
We're seeing bad timestamps in frametimeline traces (end_frame with a
lower ts than the begin event). Add a stat to track the number of such
cases per trace. There are cases where a small amount is not wrong, for
example ring buffer traces. On the other hand, numbers in the hundreds
imply that the trace has issues.

Change-Id: Ic98d5e0999417f973a1995d6d8d3d4dfa5716d88