tp: downgrade ftrace_error_stats to an info from an error statistic
Right now ~every android trace has the orange flag raised in the UI due
to atrace categories expanding to obsolete events that don't exist. So
either everyone learns to ignore the flag (potentially missing real data
loss), or gets annoyed by having to double-check whether it's
ftrace_error_stats or not.

Since we can't find the time to do anything better, just downgrade the
error. Also do it for all platforms since trace_processor doesn't
normally single out android traces.

This loses a bit of feedback for users with hand crafted configs with
e.g. a typo or a legitimate reason for an event to be missing from a
machine, but I'm saying this is an acceptable tradeoff given the effort
I'm willing to spend. The info is still there, just not an error.

Change-Id: Idf4a7751a37b8ce732c27c086deb0579e75ea6fb