perfetto: add minimal implementation of Futures
This CL adds a minimal implementation of futures in the Perfetto
codebase. This is mainly intended to be used by cloud trace processor.

Perfetto futures are inspired heavily by futures in Rust and as such
are a pull-based, lazy implementation. Unlike Rust however, the
"combinators" (i.e. the operators which allow combining futures) are
extremely minimal, only allowing the most basic operations.

Change-Id: I6ad035764b08757cea909fc15864f7dd461efc62