Add hyp_enter and hyp_exit hypervisor events and show them in UI
Hypervisor events are similar to the ftrace evetns, however they don't
have a pid. In order to work around this issue, ftrace_parser uses the
pid 0 when processing hypervisor events.

Support for showing them in the UI is quite basic so far.
Once trace_processors sees a hyp_enter event it will start a slice on
the corresponding pKVM Hypervisor CPU track. Once it sees hyp_exit
event, it will end the slice.

Bug: 249050813
Test: ui/run-dev-server
Change-Id: Ic6393bb5e11c004e0c5d0486bbe11925a4f198d0