Update Dialer HSC Query
Update the Dialer HSC query to use the last animation frame as opposed
to the first choreographer frame. Dialer appears to only play an
animation during startup.  Using when that animation finishes for the
metric brings it more inline with the physical HSC metric.

Bug: 266863423
Test: Manual testing.  Ran trace_processor_shell with the updated query
to verify script still runs and android_startup metric has changed.

Change-Id: Id7484f7f6ea21b601b1b666514e905ea14c63841